move colorscheme to lua folder
